// src/games/keybr/packages/keybr-widget/lib/components/range/Range.types.ts
import { type RefObject } from "@keybr/solid-compat/react";
import { type SizeName } from "../../styles/index.ts";
import { type Focusable, type FocusProps, type KeyboardProps, type MouseProps, } from "../types.ts";
export type RangeProps = {
    readonly max: number;
    readonly min: number;
    readonly name?: string;
    readonly ref?: RefObject<RangeRef | null>;
    readonly size?: SizeName;
    readonly step: number;
    readonly title?: string;
    readonly value: number;
    readonly onChange?: (value: number) => void;
} & FocusProps & MouseProps & KeyboardProps;
export type RangeRef = Focusable;